Rename blobpm module name
